+++ godot-client/spacetime_bindings/schema/module_blackholio_db.gd
@@ -2,19 +2,19 @@
 # FILE WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.
 class_name BlackholioModuleDb extends RefCounted
 
-const table_names : Array[StringName] = [&"circle", &"config", &"consume_entity_event", &"entity", &"food", &"player"]
+const table_names : Array[StringName] = [&"entity", &"config", &"player", &"circle", &"consume_entity_event", &"food"]
 
-var circle: BlackholioCircleTable
-var config: BlackholioConfigTable
-var consume_entity_event: BlackholioConsumeEntityEventTable
 var entity: BlackholioEntityTable
-var food: BlackholioFoodTable
+var config: BlackholioConfigTable
 var player: BlackholioPlayerTable
+var circle: BlackholioCircleTable
+var consume_entity_event: BlackholioConsumeEntityEventTable
+var food: BlackholioFoodTable
 
 func _init(p_local_db: LocalDatabase) -> void:
-	circle = preload('res://spacetime_bindings/schema/tables/blackholio_circle_table.gd').new(p_local_db)
-	config = preload('res://spacetime_bindings/schema/tables/blackholio_config_table.gd').new(p_local_db)
-	consume_entity_event = preload('res://spacetime_bindings/schema/tables/blackholio_consume_entity_event_table.gd').new(p_local_db)
 	entity = preload('res://spacetime_bindings/schema/tables/blackholio_entity_table.gd').new(p_local_db)
-	food = preload('res://spacetime_bindings/schema/tables/blackholio_food_table.gd').new(p_local_db)
+	config = preload('res://spacetime_bindings/schema/tables/blackholio_config_table.gd').new(p_local_db)
 	player = preload('res://spacetime_bindings/schema/tables/blackholio_player_table.gd').new(p_local_db)
+	circle = preload('res://spacetime_bindings/schema/tables/blackholio_circle_table.gd').new(p_local_db)
+	consume_entity_event = preload('res://spacetime_bindings/schema/tables/blackholio_consume_entity_event_table.gd').new(p_local_db)
+	food = preload('res://spacetime_bindings/schema/tables/blackholio_food_table.gd').new(p_local_db)